+++ i18n/locales/en.json
@@ -701,6 +701,16 @@
   "pages.user_profile_updated": "User profile updated",
   "pages.verified": "Verified",
   "pages.view_cart": "View Cart",
+  "plus_reading_report.listening_minutes": "Listening (min)",
+  "plus_reading_report.period": "Period",
+  "plus_reading_report.read_rate": "Reading Rate (USD/min)",
+  "plus_reading_report.reading_minutes": "Reading (min)",
+  "plus_reading_report.refresh": "Refresh",
+  "plus_reading_report.status_paid": "Paid",
+  "plus_reading_report.status_pending": "Pending",
+  "plus_reading_report.summary_total": "Total Earnings",
+  "plus_reading_report.title": "Plus Reading Library Earnings",
+  "plus_reading_report.tts_rate": "Listening Rate (USD/min)",
   "preview_book.detected_type": "Detected file type",
   "preview_book.download": "Download",
   "preview_book.error_decrypt": "Failed to decrypt book file",
