+++ i18n/locales/zh-TW.json
@@ -701,6 +701,16 @@
   "pages.user_profile_updated": "用戶資料已更新",
   "pages.verified": "已驗證",
   "pages.view_cart": "查看購物車",
+  "plus_reading_report.listening_minutes": "朗讀時間（分鐘）",
+  "plus_reading_report.period": "結算月份",
+  "plus_reading_report.read_rate": "閱讀單價（USD/分鐘）",
+  "plus_reading_report.reading_minutes": "閱讀時間（分鐘）",
+  "plus_reading_report.refresh": "重新整理",
+  "plus_reading_report.status_paid": "已發放",
+  "plus_reading_report.status_pending": "待發放",
+  "plus_reading_report.summary_total": "總收益",
+  "plus_reading_report.title": "Plus 圖書館閱覽分成報表",
+  "plus_reading_report.tts_rate": "朗讀單價（USD/分鐘）",
   "preview_book.detected_type": "偵測到的檔案類型",
   "preview_book.download": "下載",
   "preview_book.error_decrypt": "無法解密書籍檔案",
